Switched from Roboto Mono to JetBrains Mono for website's font
This is my font-of-choice for all development interfaces. It's not only clean and legible, but is open-source, made for developers, has ligature support, has italics support.

The increased letter height is especially helpful for reading sessions.

Finally, the website experiences no FOIT differences compared to using Roboto Mono. This means the website continues to have minimal to zero FOIT with fallback font optimisation.